==> src/03_ai_wiki/03_id_ner.ipynb <==
# Named Entity Recognition on AI Wiki text

Author: Crystal

Major package used: Stanza 

import pandas as pd
import numpy as np
import json
import glob
import time


# web scrapping
import requests as r
from bs4 import BeautifulSoup
import re
import bs4 as bs
import urllib

# Gensim
import gensim
import gensim.corpora as corpora
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel

#name entity recognition
import stanza
stanza.download('en') # download English model

#spacy
import spacy
import nltk
from nltk.corpus import stopwords
#spacy.load("en_core_web_sm")

#visualization
import pyLDAvis
import pyLDAvis.gensim_models as gensimvis

import warnings
#warnings.simplefilter('always')
warnings.filterwarnings("ignore", category=DeprecationWarning)

## Read in named entity data

df = pd.read_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/ai_wiki_text_entity.csv') 

df

#de-duplicate text
df_unique = df.drop_duplicates(subset="text")
len(df_unique)

df_unique["type"].value_counts(sort=True)

### Named entities that we are sure to exclude from the links (topics):
- Cardinal, just numbers, which wer are not concerned about that could be linked to another Wiki page
- Product, such as Skype, kINECT, Xobx 360
- Percent
- Ordinal
- Money
- Time
- Quantity
- Language

### Named entities that we want to examine: 


#### NER: Organization (debatable)

org_ls = df_unique[df_unique["type"] == "ORG"].text.tolist()
org_ls

#### NER: Date

date_ls = df_unique[df_unique["type"] == "DATE"].text.tolist()
date_ls

#### NER: Person

person_ls = df_unique[df_unique["type"] == "PERSON"].text.tolist()
person_ls

#### NER: WORK_OF_ART

work_of_art_ls = df_unique[df_unique["type"] == "WORK_OF_ART"].text.tolist()
work_of_art_ls

#### NER: GPE

gpe_ls = df_unique[df_unique["type"] == "GPE"].text.tolist()
gpe_ls

#### NER: LOC

loc_ls = df_unique[df_unique["type"] == "LOC"].text.tolist()
loc_ls

#### NER: NORP

norp_ls = df_unique[df_unique["type"] == "NORP"].text.tolist()
norp_ls

## Combine all the named entities

ls = org_ls + date_ls + person_ls + work_of_art_ls + gpe_ls + loc_ls  + norp_ls
