==> src/03_ai_wiki/05_scrape_ai_related_wiki.ipynb <==
# Scrape AI related Wiki articles

Author: Crystal 

import pandas as pd
import numpy as np
import json
import glob
import time


# web scrapping
import requests as r
from bs4 import BeautifulSoup
import re
import bs4 as bs
import urllib

# Gensim
import gensim
import gensim.corpora as corpora
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel

#spacy
import spacy
import nltk
from nltk.corpus import stopwords

#visualization
import pyLDAvis
import pyLDAvis.gensim_models as gensimvis

import warnings
#warnings.simplefilter('always')
warnings.filterwarnings("ignore", category=DeprecationWarning)

links_valid = pd.read_csv("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/wiki_ai_links_valid.csv") #import csv


links = links_valid["url"]

links[0]

articles = pd.DataFrame()
for link in links:
    url = "https://en.wikipedia.org/" + link
    print(url)
    source = urllib.request.urlopen(url).read()
    print("Scrapping.")
    soup = BeautifulSoup(source, 'html.parser')
    texts = ""
    for paragraph in soup.select('p'):
        texts += paragraph.text
    articles = articles.append({
        "link": link,
        "url": url,
        'article': texts
        }, ignore_index = True)

articles.to_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/wiki_ai_related_articles.csv', index = False)   

articles
